+++ antarest/study/storage/rawstudy/model/filesystem/root/input/commons/prepro_series.py
@@ -17,7 +17,7 @@ from antarest.study.storage.rawstudy.model.filesystem.common.prepro import Input
 from antarest.study.storage.rawstudy.model.filesystem.config.model import FileStudyTreeConfig
 from antarest.study.storage.rawstudy.model.filesystem.folder_node import FolderNode
 from antarest.study.storage.rawstudy.model.filesystem.inode import TREE
-from antarest.study.storage.rawstudy.model.filesystem.matrix.constants import default_scenario_hourly
+from antarest.study.storage.rawstudy.model.filesystem.matrix.simulator_default import default_scenario_hourly
 
 
 class InputPreproSeries(FolderNode):
