+++ antarest/study/storage/rawstudy/model/filesystem/root/input/hydro/prepro/area/area.py
@@ -13,8 +13,8 @@ from typing_extensions import override
 
 from antarest.study.storage.rawstudy.model.filesystem.folder_node import FolderNode
 from antarest.study.storage.rawstudy.model.filesystem.inode import TREE
-from antarest.study.storage.rawstudy.model.filesystem.matrix.constants import default_energy
 from antarest.study.storage.rawstudy.model.filesystem.matrix.input_series_matrix import InputSeriesMatrix
+from antarest.study.storage.rawstudy.model.filesystem.matrix.simulator_default import default_energy
 from antarest.study.storage.rawstudy.model.filesystem.root.input.hydro.prepro.area.prepro import (
     InputHydroPreproAreaPrepro,
 )
