+++ antarest/study/storage/rawstudy/model/filesystem/root/input/thermal/series/area/thermal/thermal.py
@@ -14,8 +14,8 @@ from typing_extensions import override
 from antarest.study.model import MatrixFrequency
 from antarest.study.storage.rawstudy.model.filesystem.folder_node import FolderNode
 from antarest.study.storage.rawstudy.model.filesystem.inode import TREE
-from antarest.study.storage.rawstudy.model.filesystem.matrix.constants import default_scenario_hourly
 from antarest.study.storage.rawstudy.model.filesystem.matrix.input_series_matrix import InputSeriesMatrix
+from antarest.study.storage.rawstudy.model.filesystem.matrix.simulator_default import default_scenario_hourly
 
 
 class InputThermalSeriesAreaThermal(FolderNode):
